pyproject.toml
[project]
name = "graphene"
version = "0.1.0"
description = "Graphene for Django"
keywords = ["graphql", "django", "python"]
authors = [
    {name = "Zadigo",email = "pendenquejohn@gmail.com"}
]
maintainers = [
    {name = "Zadigo",email = "pendenquejohn@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14,<4.0.0"
dependencies = [
    "graphql-core (>=3.2.7,<4.0.0)",
    "graphql-relay (>=3.2.0,<4.0.0)"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
    "Framework :: Django"
]

[dependency-groups]
dev = [
    "autopep8 (>=2.3.2,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "coverage (>=7.13.4,<8.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://github.com/Zadigo/graphene"
Documentation = "https://github.com/Zadigo/graphene/wiki"
Repository = "https://github.com/Zadigo/graphene.git"
Changelog = "https://github.com/Zadigo/graphene/blob/main/CHANGELOG.md"

[tool.pyright]
typeCheckingMode = "off"
extraPaths = ["./new_graphene"]
reportMissingImports = true
enableReachabilityAnalysis = true
reportUntypedBaseClass = true
